feat : 회원 탈퇴 체크 추가

+++ src/main/java/com/mzcteam01/mzcproject01be/domains/user/service/UserServiceImpl.java
@@ -254,6 +254,9 @@ public class UserServiceImpl implements UserService {
         userRepository.findById(id)
                 .orElseThrow(() -> new CustomException(UserErrorCode.USER_NOT_FOUND.getMessage()));
 
+        if( !relatedEntityChecker.relatedUserCheck( id ) )
+            throw new CustomException( CommonErrorCode.RELATED_ENTITY_EXISTED.getMessage());
+
         userRepository.deleteById(id);
     }
